+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Remove unused `copy` argument to `CellImageSequence`
 * Store reference to original for `CellImage` with the correct cell marked
 * Avoid passing geometric entities to segmented modules
+* Save image meta in separate JSON file
 
 ## 0.2.1